Trocar inteiro qtd, margem por float

+++ database/migrations/2020_05_20_150729_create_produto_venda_table.php
@@ -18,7 +18,7 @@ class CreateProdutoVendaTable extends Migration
             $table->unsignedBigInteger('produto_id');
             $table->unsignedBigInteger('venda_id');
             $table->boolean('granel');
-            $table->integer('qtd');
+            $table->float('qtd');
             $table->decimal('custo_medio', 10, 3); //histórico momento custo
             $table->decimal('valor_venda', 10, 2);
             $table->timestamps();
